Consolidate the handling of the scoped device id
