+++ endpoint-common/src/lib.rs
@@ -2,4 +2,7 @@ pub mod auth;
 pub mod command_router;
 pub mod downstream;
 pub mod error;
+mod id;
 pub mod x509;
+
+pub use id::*;
